requirements.txt
# Core dependencies
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
scipy>=1.7.0

# Deep learning
torch>=1.10.0
torch-geometric>=2.0.0

# Gradient boosting
xgboost>=1.5.0

# Visualization
matplotlib>=3.4.0
seaborn>=0.11.0

# Explainability
shap>=0.40.0

# Data processing
pyarrow>=6.0.0
tqdm>=4.62.0

# Utilities
pyyaml>=5.4.0
jsonlines>=2.0.0
